+++ changelog.md
@@ -4,6 +4,7 @@
 - Bump hahomematic to 2025.7.0
   - Add default customization for ELV-SH-SW1-BAT (Deleting of obsolete entities might be required)
   - Enable OPERATING_VOLTAGE_LEVEL for HM-CC-RT-DN and HM-TC-IT-WM-W-EU
+- Check uniqueness of instance name
 
 # Version 1.84.1 (2025-06-27)